# apps/server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Atarimae API",
    "version": "0.0.0",
    "description": "Self-hosted communication board for small teams.\n\nAll timestamps are UTC ISO 8601. All identifiers are UUIDv7.",
    "license": {
      "name": "AGPL-3.0-only",
      "url": "https://www.gnu.org/licenses/agpl-3.0.html"
    }
  },
...
  "paths": {
...
    "/auth/login": {
      "post": {
        "summary": "Sign in",
        "tags": [
          "auth"
        ],
        "description": "An account may be signed in on any number of devices at once. Signing in on a second device never invalidates the first.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "format": "email",
                    "maxLength": 254,
                    "description": "Case-insensitive. Stored as entered, compared in lower case.",
                    "type": "string"
                  },
                  "password": {
                    "minLength": 1,
                    "maxLength": 512,
                    "type": "string"
                  },
                  "deviceToken": {
                    "minLength": 16,
                    "maxLength": 128,
                    "type": "string"
                  },
                  "deviceName": {
                    "maxLength": 100,
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "user"
                  ],
                  "properties": {
                    "user": {
                      "type": "object",
                      "required": [
                        "id",
                        "email",
                        "displayName",
                        "role",
                        "createdAt",
                        "lastLoginAt"
                      ],
                      "properties": {
                        "id": {
                          "format": "uuid",
                          "description": "UUIDv7 identifier.",
                          "type": "string"
                        },
                        "email": {
                          "format": "email",
                          "maxLength": 254,
                          "description": "Case-insensitive. Stored as entered, compared in lower case.",
                          "type": "string"
                        },
                        "displayName": {
                          "minLength": 1,
                          "maxLength": 100,
                          "type": "string"
                        },
                        "role": {
                          "description": "owner > admin > member.",
                          "anyOf": [
                            {
                              "type": "string",
                              "enum": [
                                "owner"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "admin"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "member"
                              ]
                            }
                          ]
                        },
                        "createdAt": {
                          "format": "date-time",
                          "description": "UTC timestamp, ISO 8601.",
                          "type": "string"
                        },
                        "lastLoginAt": {
                          "anyOf": [
                            {
                              "format": "date-time",
                              "description": "UTC timestamp, ISO 8601.",
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "message"
                  ],
                  "properties": {
                    "code": {
                      "description": "Stable machine-readable error code.",
                      "type": "string",
                      "example": "NO_ELIGIBLE_RECIPIENTS"
                    },
                    "message": {
                      "description": "Developer-facing English description. Not shown to end users.",
                      "type": "string"
                    },
                    "details": {
                      "description": "Structured context. Never contains secrets.",
                      "type": "object",
                      "additionalProperties": {}
                    },
                    "requestId": {
                      "description": "Correlates with server logs and audit_logs.",
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "message"
                  ],
                  "properties": {
                    "code": {
                      "description": "Stable machine-readable error code.",
                      "type": "string",
                      "example": "NO_ELIGIBLE_RECIPIENTS"
                    },
                    "message": {
                      "description": "Developer-facing English description. Not shown to end users.",
                      "type": "string"
                    },
                    "details": {
                      "description": "Structured context. Never contains secrets.",
                      "type": "object",
                      "additionalProperties": {}
                    },
                    "requestId": {
                      "description": "Correlates with server logs and audit_logs.",
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "message"
                  ],
                  "properties": {
                    "code": {
                      "description": "Stable machine-readable error code.",
                      "type": "string",
                      "example": "NO_ELIGIBLE_RECIPIENTS"
                    },
                    "message": {
                      "description": "Developer-facing English description. Not shown to end users.",
                      "type": "string"
                    },
                    "details": {
                      "description": "Structured context. Never contains secrets.",
                      "type": "object",
                      "additionalProperties": {}
                    },
                    "requestId": {
                      "description": "Correlates with server logs and audit_logs.",
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "message"
                  ],
                  "properties": {
                    "code": {
                      "description": "Stable machine-readable error code.",
                      "type": "string",
                      "example": "NO_ELIGIBLE_RECIPIENTS"
                    },
                    "message": {
                      "description": "Developer-facing English description. Not shown to end users.",
                      "type": "string"
                    },
                    "details": {
                      "description": "Structured context. Never contains secrets.",
                      "type": "object",
                      "additionalProperties": {}
                    },
                    "requestId": {
                      "description": "Correlates with server logs and audit_logs.",
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "message"
                  ],
                  "properties": {
                    "code": {
                      "description": "Stable machine-readable error code.",
                      "type": "string",
                      "example": "NO_ELIGIBLE_RECIPIENTS"
                    },
                    "message": {
                      "description": "Developer-facing English description. Not shown to end users.",
                      "type": "string"
                    },
                    "details": {
                      "description": "Structured context. Never contains secrets.",
                      "type": "object",
                      "additionalProperties": {}
                    },
                    "requestId": {
                      "description": "Correlates with server logs and audit_logs.",
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "413": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "message"
                  ],
                  "properties": {
                    "code": {
                      "description": "Stable machine-readable error code.",
                      "type": "string",
                      "example": "NO_ELIGIBLE_RECIPIENTS"
                    },
                    "message": {
                      "description": "Developer-facing English description. Not shown to end users.",
                      "type": "string"
                    },
                    "details": {
                      "description": "Structured context. Never contains secrets.",
                      "type": "object",
                      "additionalProperties": {}
                    },
                    "requestId": {
                      "description": "Correlates with server logs and audit_logs.",
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "415": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "message"
                  ],
                  "properties": {
                    "code": {
                      "description": "Stable machine-readable error code.",
                      "type": "string",
                      "example": "NO_ELIGIBLE_RECIPIENTS"
                    },
                    "message": {
                      "description": "Developer-facing English description. Not shown to end users.",
                      "type": "string"
                    },
                    "details": {
                      "description": "Structured context. Never contains secrets.",
                      "type": "object",
                      "additionalProperties": {}
                    },
                    "requestId": {
                      "description": "Correlates with server logs and audit_logs.",
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "message"
                  ],
                  "properties": {
                    "code": {
                      "description": "Stable machine-readable error code.",
                      "type": "string",
                      "example": "NO_ELIGIBLE_RECIPIENTS"
                    },
                    "message": {
                      "description": "Developer-facing English description. Not shown to end users.",
                      "type": "string"
                    },
                    "details": {
                      "description": "Structured context. Never contains secrets.",
                      "type": "object",
                      "additionalProperties": {}
                    },
                    "requestId": {
                      "description": "Correlates with server logs and audit_logs.",
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "message"
                  ],
                  "properties": {
                    "code": {
                      "description": "Stable machine-readable error code.",
                      "type": "string",
                      "example": "NO_ELIGIBLE_RECIPIENTS"
                    },
                    "message": {
                      "description": "Developer-facing English description. Not shown to end users.",
                      "type": "string"
                    },
                    "details": {
                      "description": "Structured context. Never contains secrets.",
                      "type": "object",
                      "additionalProperties": {}
                    },
                    "requestId": {
                      "description": "Correlates with server logs and audit_logs.",
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "message"
                  ],
                  "properties": {
                    "code": {
                      "description": "Stable machine-readable error code.",
                      "type": "string",
                      "example": "NO_ELIGIBLE_RECIPIENTS"
                    },
                    "message": {
                      "description": "Developer-facing English description. Not shown to end users.",
                      "type": "string"
                    },
                    "details": {
                      "description": "Structured context. Never contains secrets.",
                      "type": "object",
                      "additionalProperties": {}
                    },
                    "requestId": {
                      "description": "Correlates with server logs and audit_logs.",
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Version 1"
    }
  ],
...
}